feat: generate selected libraries
